Fix dom loader
